Round-1A/requirements.txt
pandas
scikit-learn
pdfminer.six
joblib
